feat(dag): modularise dag
License: MIT
